pyproject.toml
[project]
name = 'get-ip'
version = '0.1.0'
description = ''
authors = [
    {name = 'AzureSky',email = 'esipov54321@gmail.com'}
]
readme = 'README.md'
requires-python = '>=3.12'
dependencies = [
    'curl-cffi (>=0.13.0,<0.14.0)',
    'fastapi (>=0.120.4,<0.121.0)',
    'uvicorn (>=0.38.0,<0.39.0)',
    'python-dotenv (>=1.2.1,<2.0.0)',
    'sqlalchemy[asyncio] (>=2.0.44,<3.0.0)',
    'asyncpg (>=0.31.0,<0.32.0)',
    'alembic (>=1.17.2,<2.0.0)',
    'pydantic-settings (>=2.12.0,<3.0.0)',
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ['poetry-core>=2.0.0,<3.0.0']
build-backend = 'poetry.core.masonry.api'

[dependency-groups]
dev = [
    'black (>=25.9.0,<26.0.0)',
    'isort (>=7.0.0,<8.0.0)',
    'ruff (>=0.14.3,<0.15.0)',
    'mypy (>=1.18.2,<2.0.0)'
]

[tool.black]
skip-string-normalization = true
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
